src/slices/user/controllers/deleteUser/deleteUserController.ts
/* eslint-disable no-unsafe-optional-chaining */
import {
  HttpRequest,
  HttpResponse,
  Validation,
  badRequest,
  ok,
} from "@/application/helpers";
import { Controller } from "@/application/infra/contracts";
import { DeleteUser } from "@/slices/user/useCases";

export class DeleteUserController extends Controller {
  constructor(
    private readonly validation: Validation,
    private readonly deleteUser: DeleteUser
  ) {
    super();
  }
  async execute(httpRequest: HttpRequest<any>): Promise<HttpResponse<any>> {
    const errors = this.validation.validate(httpRequest?.query);
    if (errors?.length > 0) {
      return badRequest(errors);
    }
    const userDeleteed = await this.deleteUser({
      fields: { ...httpRequest?.query, createdById: httpRequest?.userId },
      options: {},
    });
    return ok(userDeleteed);
  }
}
